foodgram/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-filter==23.2
djoser
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
psycopg2-binary==2.9.7

